+++ aimage/abc.py
@@ -1,6 +1,5 @@
 from abc import ABC
 
-import discord
 from aiohttp import ClientSession
 from redbot.core import Config, commands
 from redbot.core.bot import Red
@@ -18,17 +17,6 @@ class MixinMeta(ABC):
     autocomplete_cache: dict
 
     def __init__(self, *args):
-        pass
-
-    async def _get_endpoint(self, guild: discord.Guild):
-        """ Gets the correct endpoint for the guild """
-        pass
-
-    async def _fetch_data(self, guild: discord.Guild, endpoint_suffix: str):
-        """ Helper function to fetch data from Stable Diffusion endpoint """
-        pass
-
-    async def get_auth(self, auth_str: str):
         pass
 
     async def generate_image(self, *args, **kwargs):
